requirements.txt
langchain==0.2.12
langchain-community==0.2.11
langchain-ollama==0.1.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
duckduckgo-search==6.2.6
transformers==4.43.3
torch>=2.2.0
python-dotenv==1.0.1
rank_bm25==0.2.2
fastapi==0.111.0
uvicorn==0.30.1
pypdf==4.3.1
markdown2==2.5.0
